License: update licensing to GNU GPLv2 or later
The update is done in order to comply with both the licensing of some
code snippets being used, and our intentions.

This program is free software; you can redistribute it and/or modify
it under the terms of the GNU General Public License as published by
the Free Software Foundation; either version 2 of the License, or
(at your option) any later version.

Signed-off-by: Andrey Gelman <andrey.gelman@compulab.co.il>