options: add version reporting
The daemon should be able to report its version so we could tell what binary
we are dealing with.

Implement the --version flag, and setup the build system such that version
will be set in Makefile, and appanded with information about the state of the
git repository.

This project shall adhere to Semantic Versioning.

Signed-off-by: Nikita Kiryanov <nikita@compulab.co.il>